+++ packages/forms/src/UIForm/fields/Enumeration/EnumerationWidget.js
@@ -49,7 +49,12 @@ class EnumerationForm extends React.Component {
 		return ITEMS_DEFAULT_HEIGHT;
 	}
 
-	static parseStringValueToArray(values) {
+	static parseStringValueToArray(values, skipCommas) {
+		if (skipCommas) {
+			return values
+				.match(/(\\.|[^,])+/g)
+				.map(value => value.trim().replace(/\\,/g, ',').replace(/\\./g, '\\'));
+		}
 		return values.split(',').map(value => value.trim());
 	}
 
@@ -432,7 +437,10 @@ class EnumerationForm extends React.Component {
 					actionsEdit: this.loadingInputsActions,
 				},
 			}));
-			const formattedValue = EnumerationForm.parseStringValueToArray(value.value);
+			const formattedValue = EnumerationForm.parseStringValueToArray(
+				value.value,
+				this.props.skipCommas,
+			);
 			this.props
 				.onTrigger(event, {
 					trigger: {
@@ -467,9 +475,10 @@ class EnumerationForm extends React.Component {
 			}
 			item.displayMode = enumerationStates.DISPLAY_MODE_DEFAULT;
 			const valueExist = this.valueAlreadyExist(value.value, this.state);
+
 			// if the value is empty, no value update is done
 			if (value.value && !valueExist) {
-				item.values = EnumerationForm.parseStringValueToArray(value.value);
+				item.values = EnumerationForm.parseStringValueToArray(value.value, this.props.skipCommas);
 			}
 			if (valueExist) {
 				item.error = this.props.t('ENUMERATION_WIDGET_DUPLICATION_ERROR', {
@@ -698,10 +707,11 @@ class EnumerationForm extends React.Component {
 			this.setState({
 				headerInput: this.loadingInputsActions,
 			});
+
 			this.props
 				.onTrigger(event, {
 					trigger: {
-						value: EnumerationForm.parseStringValueToArray(value.value),
+						value: EnumerationForm.parseStringValueToArray(value.value, this.props.skipCommas),
 						action: ENUMERATION_ADD_ACTION,
 					},
 					schema,
@@ -725,7 +735,7 @@ class EnumerationForm extends React.Component {
 				schema,
 				value: this.state.items.concat([
 					{
-						values: EnumerationForm.parseStringValueToArray(value.value),
+						values: EnumerationForm.parseStringValueToArray(value.value, this.props.skipCommas),
 					},
 				]),
 			};
@@ -1050,6 +1060,7 @@ if (process.env.NODE_ENV !== 'production') {
 		onTrigger: PropTypes.func.isRequired,
 		properties: PropTypes.object,
 		schema: PropTypes.object,
+		skipCommas: PropTypes.bool,
 		t: PropTypes.func,
 		value: PropTypes.arrayOf(
 			PropTypes.shape({