+++ tools/scripts-config-jest/jest.config.js
@@ -1,9 +1,5 @@
 const path = require('path');
-const { jsWithBabel: tsjestPreset } = require('ts-jest/presets');
-const {
-	getBabelConfig,
-	getBabelConfigPath,
-} = require('@talend/scripts-config-babel/babel-resolver');
+const { getBabelConfigPath } = require('@talend/scripts-config-babel/babel-resolver');
 
 module.exports = {
 	moduleNameMapper: {
@@ -12,19 +8,13 @@ module.exports = {
 	},
 	rootDir: process.cwd(),
 	setupFilesAfterEnv: [path.join(__dirname, 'test-setup.js')],
-	testEnvironment: 'jsdom',
+	testEnvironment: 'jest-environment-jsdom-global',
 	testRegex: 'src/.*\\.test.(js|ts|tsx)$',
 	transform: {
-		...tsjestPreset.transform, // https://kulshekhar.github.io/ts-jest/user/config/#advanced
 		'^.+\\.jsx?$': ['babel-jest', { configFile: getBabelConfigPath() }],
 		'^.+\\.mjs?$': ['babel-jest', { configFile: getBabelConfigPath() }],
+		'^.+\\.tsx?$': ['babel-jest', { configFile: getBabelConfigPath() }],
 	},
 	snapshotSerializers: ['jest-serializer-html'],
 	modulePathIgnorePatterns: ['<rootDir>/dist/cdn'],
-	globals: {
-		'ts-jest': {
-			// https://kulshekhar.github.io/ts-jest/user/config/babelConfig
-			babelConfig: getBabelConfig(),
-		},
-	},
 };