+++ tools/scripts-config-react-webpack/config/webpack.config.common.js
@@ -23,7 +23,7 @@ function getSassData(userSassData) {
 }
 
 function getCommonStyleLoaders(enableModules, mode) {
-	const sourceMap = mode === 'production';
+	const sourceMap = true;
 	let cssOptions = {
 		sourceMap,
 	};
@@ -37,9 +37,7 @@ function getCommonStyleLoaders(enableModules, mode) {
 		};
 	}
 	return [
-		mode === 'development'
-			? { loader: 'style-loader' }
-			: { loader: MiniCssExtractPlugin.loader, options: { esModule: false } },
+		{ loader: MiniCssExtractPlugin.loader, options: { esModule: false } },
 		{ loader: 'css-loader', options: cssOptions },
 		{
 			loader: 'postcss-loader',
@@ -64,7 +62,7 @@ function getJSAndTSLoader(env, useTypescript) {
 }
 
 function getSassLoaders(enableModules, sassData, mode) {
-	const sourceMap = mode === 'production';
+	const sourceMap = true;
 	return getCommonStyleLoaders(enableModules, mode).concat(
 		{ loader: 'resolve-url-loader', options: { sourceMap } },
 		{