+++ packages/components/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @talend/react-components
 
+## 7.6.0
+
+### Minor Changes
+
+- eba1ac047: feat: add HeaderBar and AppSwitcher icon props
+
+## 7.5.0
+
+### Minor Changes
+
+- a99154a7d: generate minified css using dot min in the name of it
+
+### Patch Changes
+
+- Updated dependencies [617ec14f0]
+- Updated dependencies [a99154a7d]
+  - @talend/design-system@4.0.0
+
 ## 7.4.0
 
 ### Minor Changes