+++ tools/scripts-config-jest/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @talend/scripts-config-jest
 
+## 11.2.1
+
+### Patch Changes
+
+- c815ba11f: fix: use babel to transpile ts[x] tests
+
+## 11.2.0
+
+### Minor Changes
+
+- ab1a6e983: feat: mock ResizeObserver
+
+### Patch Changes
+
+- 4d8636214: expose jsdom as global variable in jest environment
+
 ## 11.1.0
 
 ### Minor Changes